Move HomeController & Fix other controller
