+++ app/Http/Controllers/ProductController.php
@@ -4,13 +4,19 @@ namespace App\Http\Controllers;
 
 use App\Models\Product;
 use App\Models\ErrorLog;
+use App\Models\Category;
 use App\Helpers\StringHelper;
+use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use App\Utilities\GoogleDriveUtility;
+use App\Utilities\ProductsUtility;
+use App\Interfaces\StatusInterface;
+use App\Http\Controllers\Controller;
 
-class ProductController extends Controller
+class ProductController extends Controller implements StatusInterface
 {
     private $googleDriveUtility;
-    private $productsUtility;
+    private $productUtility;
 
     /**
      * Summary of __construct
@@ -18,6 +24,7 @@ class ProductController extends Controller
     public function __construct()
     {
         $this->googleDriveUtility = new GoogleDriveUtility();
+        $this->productUtility = new ProductsUtility();
     }
 
     /**
@@ -57,5 +64,29 @@ class ProductController extends Controller
         ];
 
         return view('product', compact('product', 'productImgs'));
+    }
+
+    /**
+     * Summary of sortProducts
+     * @param \Illuminate\Http\Request $request
+     * @param string $category
+     * @param string $sort
+     * @return mixed|\Illuminate\Http\JsonResponse
+     */
+    public function sortProducts(Request $request, $category, $sort)
+    {
+        if (!$request->ajax()) abort(404);
+
+        $category = Category::where('name', $category)->first();
+
+        if (!$category) abort(404);
+
+        $products = $this->productUtility->getProducts($category->id, (int) $sort);
+        
+        return response()->json([
+            'status' => self::STATUS_SUCCESS,
+            'message' => 'Data sorted!',
+            'data' => $products,
+        ], Response::HTTP_OK);
     }
 }
