+++ routes/api.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Http\Controllers\ProductsController;
+use App\Http\Controllers\ProductController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -19,4 +19,4 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::get('/category/{category}/{sort}', [ProductsController::class, 'sortProducts'])->name('sortProducts');
+Route::get('/category/{category}/{sort}', [ProductController::class, 'sortProducts'])->name('sortProducts');
